Make it easy to start RabbitMQ from 3.6.x
While answering "Can RabbitMQ 3.6 can be upgraded to 3.8 in-place?" it
was necessary to start the server from 3.6. These changes enabled me to
use the following command:

    # Start as RabbitMQ v3.6.16
    make server DOCKER_RABBITMQ_IMAGE=rabbitmq:3.6.16-management
    # Upgrade to RabbitMQ v3.8.2
    make server-38x

I only did a single-node, and the upgrade went through without a hitch
in 3' 47". Out of that, 2' of that was GCP "deleting" the VM and another
1' was booting the new VM, puling the rabbitmq:3.8.2-management image
etc.

re https://groups.google.com/d/msg/rabbitmq-users/DnFmA54_LJk/-jvRAddRAgAJ

[#170544206]

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>