Install kube-prometheus using the official docs
There seems to be a permissions issue in k3s:

    Failed to list *v1.Service: services is forbidden: User \"system:serviceaccount:monitoring:prometheus-k8s\" cannot list resource \"services\" in API group \"\" in the namespace \"default\""

I found this which seems to be a simpler way of getting Prometheus &
Grafana onto k3s: https://github.com/cablespaghetti/k3s-monitoring

Going to try that next as the current approach feels like a rabbit hole
based on what we care about. Committing it to capture what we have so
far, even though it may end up discarded.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>