This property MUST be configured if other similar properties are
Otherwise RabbitMQ will fail to boot due to missing
sample_retention_policies. By the way, this property is not used because
we don't have detailed rates_mode set, but there must be a map merge
happening in Cuttlefish that results in an incomplete config which
RabbitMQ fails to handle and crashes.

cc @lukebakken @michaelklishin

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>